% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_networks.R
\name{add_base_effect_gaussian}
\alias{add_base_effect_gaussian}
\title{Add a base effect to all edges to the GRN
At the moment this samples from a random Gaussian with mean 2.5 and standard deviation 1}
\usage{
add_base_effect_gaussian(net, disregulated_regulators, mean = 2.5, sd = 1)
}
\arguments{
\item{net}{The network as an edge list}

\item{disregulated_regulators}{The data table containing the perturbation information}
}
\value{
The network with an additional column base_effect containing the base weights
}
\description{
Add a base effect to all edges to the GRN
At the moment this samples from a random Gaussian with mean 2.5 and standard deviation 1
}
