% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_networks.R
\name{add_base_effect_to_grn}
\alias{add_base_effect_to_grn}
\title{Title}
\usage{
add_base_effect_to_grn(net, disregulated_regulators)
}
\arguments{
\item{disregulated_regulators}{}
}
\description{
Title
}
