% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_networks.R
\name{batch_create_gex_data}
\alias{batch_create_gex_data}
\title{Title}
\usage{
batch_create_gex_data(
  net.dir,
  gex.dir,
  network.list,
  nr.modules = 2,
  nr.genes.per.module = 2,
  disregulation.type = "tf-effect-0",
  seed = 11
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
