% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_modules.R
\name{cluster}
\alias{cluster}
\title{Recursively cluster and save the modules}
\usage{
cluster(gr, min_nodes = 50, max_nodes = 200)
}
\arguments{
\item{gr}{Current subgraph}

\item{network_dir}{Output dir (trailed along)}

\item{orig_graph}{Original graph (trailed along)}
}
\description{
Recursively cluster and save the modules
}
