% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_modules.R
\name{cluster_and_plot}
\alias{cluster_and_plot}
\title{Use spectral clustering to partition the graph into n.cluster}
\usage{
cluster_and_plot(edgelist, basedir, n.clusters = 3)
}
\arguments{
\item{edgelist}{file with list of network files}

\item{basedir}{Directory}

\item{n.clusters}{number of clusters to produce}
}
\description{
Use spectral clustering to partition the graph into n.cluster
}
