% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_networks.R
\name{create_gex_data}
\alias{create_gex_data}
\title{Generate data for a list of pregenerated networks.}
\usage{
create_gex_data(
  net,
  node_labels,
  net_name,
  nr_modules = 2,
  nr_genes_per_module = 2,
  base_effect = "standard-normal",
  mean = 2.5,
  sd = 1,
  disregulation_type = "remove-tf-effect-others",
  seed = 11,
  tf_effect = 5.5,
  weight_delta = 0.5
)
}
\arguments{
\item{seed}{}
}
\description{
Generate data for a list of pregenerated networks.
}
