% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_networks.R
\name{create_gex_data_easy}
\alias{create_gex_data_easy}
\title{Generate data for a list of pregenerated networks.}
\usage{
create_gex_data_easy(
  net,
  common_net,
  net_name,
  base_effect = "standard-normal",
  mean = 2.5,
  sd = 1,
  seed = 11
)
}
\arguments{
\item{net_name}{}
}
\description{
Generate data for a list of pregenerated networks.
}
