% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_modules.R
\name{createGraph}
\alias{createGraph}
\title{Create Graph from data.table.}
\usage{
createGraph(network_dt, directed = FALSE)
}
\arguments{
\item{network_dt}{A data.table containing at least a colum 'source' and 'target' as the first and second column}

\item{directed}{set to true if the igraph object should contain directed edges (no greedy clustering possible)}
}
\value{
An igraph object without multi-egdes
}
\description{
Create Graph from data.table.
}
\examples{
collectri_file <- '/tmp/collectri.tsv'
collectri <- loadOrDownloadCollectTRI(collectri_file)
colnames(collectri)<- c('source', 'target', 'weight')
orig_graph<-createGraph(collectri)
}
