% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_networks.R
\name{generate_data_from_grn}
\alias{generate_data_from_grn}
\title{Title}
\usage{
generate_data_from_grn(
  net,
  n.cells = 750,
  seed = 11,
  tree = Phyla1(),
  noise = FALSE
)
}
\arguments{
\item{noise}{}
}
\description{
Title
}
