% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn2gex-package.R
\docType{package}
\name{grn2gex-package}
\alias{grn2gex}
\alias{grn2gex-package}
\title{grn2gex: Generate Differential Networks and Gene Expession data from those Networks}
\description{
This package creates differential gene regulatory networks from an existing structure. The networks are then used to generate differential gene expression data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AnneHartebrodt/grn2gex}
  \item Report bugs at \url{https://github.com/AnneHartebrodt/grn2gex/issues}
}

}
\author{
\strong{Maintainer}: Anne Hartebrodt \email{anne.hartebrodt@fau.de} (\href{https://orcid.org/0000-0002-9172-3137}{ORCID})

}
\keyword{internal}
