% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_modules.R
\name{loadOrDownloadCollectTRI}
\alias{loadOrDownloadCollectTRI}
\title{Convenience function to either use an existing collectri file or download it from the web via DecouplR.}
\usage{
loadOrDownloadCollectTRI(collectri_file, organism = "human")
}
\arguments{
\item{collectri_file}{Filename of the Collectri file, if exists will be used, otherwise downloaded via decoupleR and saved}
}
\value{
A data.table with the network of the specified organism
}
\description{
Convenience function to either use an existing collectri file or download it from the web via DecouplR.
}
\examples{
collectri_file <- '/tmp/collectri.tsv'
collectri <- loadOrDownloadCollectTRI(collectri_file)
}
