% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_networks.R
\name{modify_weight_of_other_edges}
\alias{modify_weight_of_other_edges}
\title{Modifies the selected TF in the GRNS which are NOT part of the selected module.
Good for reducing the effect of TFs in all but the selected module}
\usage{
modify_weight_of_other_edges(net, disregulated_regulators, weight_delta = 0.5)
}
\arguments{
\item{net}{The input network with columns source, target and base_effect}

\item{disregulated_regulators}{The mapping with the perturbed regulators}

\item{weight_delta}{The weight to be added/substracted}
}
\value{
net A network with an additional column grn_effect which is a copy of the base effect but modified according
to the network mapping
}
\description{
Modifies the selected TF in the GRNS which are NOT part of the selected module.
Good for reducing the effect of TFs in all but the selected module
}
