% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_modules.R
\name{remove_irrelevant_networks}
\alias{remove_irrelevant_networks}
\title{Remove networks which contain only one or 2 hubnodes}
\usage{
remove_irrelevant_networks(basedir)
}
\arguments{
\item{basedir}{Directory containing network files with the ending ".edges.tsv"}
}
\value{
edgelist: A list of files relative to the basedir containing networks with more than 1 hubnode
}
\description{
Remove networks which contain only one or 2 hubnodes
}
