% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_modules.R
\name{restore_original_directionality}
\alias{restore_original_directionality}
\title{Subcluster a small graph and save the results to file}
\usage{
restore_original_directionality(sub, orig_graph)
}
\arguments{
\item{sub}{Subgraph to be clustered and saved}

\item{orig_graph}{The original graph which is needed to recover the edge directions}

\item{network_dir}{Directory where nodes.tsv, edges.tsv, net.pdf will be saved (filenames autogenerated)}
}
\value{
A list of objects
}
\description{
Subcluster a small graph and save the results to file
}
