% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_generated_data}
\alias{save_generated_data}
\title{Save everything to a specified location}
\usage{
save_generated_data(net, counts, meta, gex_dir, disregulated_info = NULL)
}
\arguments{
\item{seed}{}
}
\description{
Save everything to a specified location
}
