% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{save_plots}
\alias{save_plots}
\title{Save diagnostic plots with the data}
\usage{
save_plots(counts, meta, directory)
}
\arguments{
\item{counts}{count matrix!! (not data.table)}

\item{meta}{metadata data.table}

\item{directory}{directory where to save the plots to}
}
\description{
Save diagnostic plots with the data
}
