% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_modules.R
\name{save_small_subgraph}
\alias{save_small_subgraph}
\title{Subcluster a small graph and save the results to file}
\usage{
save_small_subgraph(sub, network.dir, orig.graph)
}
\arguments{
\item{sub}{Subgraph to be clustered and saved}

\item{network.dir}{Directory where nodes.tsv, edges.tsv, net.pdf will be saved (filenames autogenerated)}

\item{orig.graph}{The original graph which is needed to recover the edge directions}
}
\description{
Subcluster a small graph and save the results to file
}
