% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_modules.R
\name{subsample_and_sparsify_network}
\alias{subsample_and_sparsify_network}
\title{Convenience function to either use an existing collectri file or download it from the web via DecouplR.}
\usage{
subsample_and_sparsify_network(network, num_sources, max_out_degree)
}
\arguments{
\item{network.}{A data table with source and target columns}
}
\value{
A data.table with the data table of the subnetwork
}
\description{
Convenience function to either use an existing collectri file or download it from the web via DecouplR.
}
