% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_modules.R
\name{cluster}
\alias{cluster}
\title{Recursively cluster and save the modules}
\usage{
cluster(gr, network.dir, orig.graph)
}
\arguments{
\item{gr}{Current subgraph}

\item{network.dir}{Output dir (trailed along)}

\item{orig.graph}{Original graph (trailed along)}
}
\description{
Recursively cluster and save the modules
}
