% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_networks.R
\name{modify_weight_of_other_edges}
\alias{modify_weight_of_other_edges}
\title{Title}
\usage{
modify_weight_of_other_edges(net, disregulated_regulators, weight_delta = 0.5)
}
\arguments{
\item{weight_delta}{}
}
\description{
Title
}
