% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_networks.R
\name{remove_master_tf_effect_from_other_modules}
\alias{remove_master_tf_effect_from_other_modules}
\title{Title}
\usage{
remove_master_tf_effect_from_other_modules(net, disregulated_regulators)
}
\arguments{
\item{disregulated_regulators}{}
}
\description{
Title
}
