% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_small_subgraph}
\alias{save_small_subgraph}
\title{Subcluster a small graph and save the results to file}
\usage{
save_small_subgraph(sub, network_dir, orig_graph)
}
\arguments{
\item{sub}{Subgraph to be clustered and saved}

\item{network_dir}{Directory where nodes.tsv, edges.tsv, net.pdf will be saved (filenames autogenerated)}

\item{orig_graph}{The original graph which is needed to recover the edge directions}
}
\description{
Subcluster a small graph and save the results to file
}
\examples{
collectri.file <- '/tmp/collectri.tsv'
net.dir <- '/tmp/subgraphs'
collectri <- loadOrDownloadCollectTRI(collectri.file)
colnames(collectri)<- c('source', 'target', 'sad')
graph_list<-clusterNetwork(collectri )
graph_list<-remove_irrelevant_networks(graph_list)
orig_graph<-createGraph(collectri)
for (g in graph_list){
 save_small_subgraph(g, net.dir, orig_graph)
}
}
