% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_networks.R
\name{set_master_tf_effect}
\alias{set_master_tf_effect}
\title{Title}
\usage{
set_master_tf_effect(net, disregulated_regulators, tf_effect = 5)
}
\arguments{
\item{tf_effect}{}
}
\description{
Title
}
